requirements.txt
flask-cors
Flask-SQLAlchemy
Flask-JWT
